feat(): Adding aliases to service export
There is a requirement to identify exported services with arbitrary names
instead of slice.local names. This can be achieved by providing an alias field
in the service export spec section so that the users can give any name that
they desire to the service being exported from a cluster.

Added a new field to hold alias names to the service import and export CRDs.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>