fix(): Changed istio serviceentry resolution to static
Changed the istio serviceentry resolution to static from dns when
setting up the istio egress gw. There is a problem with dns resolution
when the istio-proxy is run in the router mode wherein the resolution of
hostnames fails intermittently and in some cases it remains in the
failed state for a very long time. This issue is not seen with
istio-proxy running in sidecar mode.

We use static resolution for istio ingress gw, following the same
practice for egress gw as well.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>