fix(): Fixed Slice Gw reconciliation issues (#428)
* fix(): Avoid unnecessary reconcile cycles for slice gw

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>

* fix(): Fixed Slice Gw reconciliation issues

Fixed the following issues:

The slicegw reconciler does not update all gw deployments when there
is a mismatch in the sidecar image.

The slicegw reconciler does not send connection ctx to all the gw pods
when one of the gw pods is not healthy.

The gw nodeport services are reconciled independently of the gw deployments
causing a mismatch in the gw services and the corresponding backends.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>

---------

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>
Signed-off-by: kranthi <kranthi@aveshasystems.com>